+++ test/Twilio.Test/TwiML/ConnectTest.cs
@@ -94,6 +94,44 @@ namespace Twilio.Tests.TwiML
                 new[] {Conversation.EventEnum.CallInitiated}
             );
 
+            elem.ConversationRelay(
+                "url",
+                "language",
+                "tts_language",
+                "transcription_language",
+                "tts_provider",
+                "voice",
+                "transcription_provider",
+                "speech_model",
+                true,
+                true,
+                "welcome_greeting",
+                true,
+                true,
+                true,
+                true,
+                true
+            );
+
+            elem.Assistant(
+                "id",
+                "language",
+                "tts_language",
+                "transcription_language",
+                "tts_provider",
+                "voice",
+                "transcription_provider",
+                "speech_model",
+                true,
+                true,
+                "welcome_greeting",
+                true,
+                true,
+                true,
+                true,
+                true
+            );
+
             Assert.AreEqual(
                 "<?xml version=\"1.0\" encoding=\"utf-8\"?>" + Environment.NewLine +
                 "<Connect>" + Environment.NewLine +
@@ -102,6 +140,8 @@ namespace Twilio.Tests.TwiML
                 "  <Stream name=\"name\" connectorName=\"connector_name\" url=\"url\" track=\"inbound_track\" statusCallback=\"status_callback\" statusCallbackMethod=\"GET\"></Stream>" + Environment.NewLine +
                 "  <VirtualAgent connectorName=\"connector_name\" language=\"language\" sentimentAnalysis=\"true\" statusCallback=\"status_callback\" statusCallbackMethod=\"GET\"></VirtualAgent>" + Environment.NewLine +
                 "  <Conversation serviceInstanceSid=\"service_instance_sid\" inboundAutocreation=\"true\" routingAssignmentTimeout=\"1\" inboundTimeout=\"1\" url=\"https://example.com\" method=\"GET\" record=\"do-not-record\" trim=\"trim-silence\" recordingStatusCallback=\"https://example.com\" recordingStatusCallbackMethod=\"GET\" recordingStatusCallbackEvent=\"in-progress\" statusCallback=\"https://example.com\" statusCallbackMethod=\"GET\" statusCallbackEvent=\"call-initiated\"></Conversation>" + Environment.NewLine +
+                "  <ConversationRelay url=\"url\" language=\"language\" ttsLanguage=\"tts_language\" transcriptionLanguage=\"transcription_language\" ttsProvider=\"tts_provider\" voice=\"voice\" transcriptionProvider=\"transcription_provider\" speechModel=\"speech_model\" profanityFilter=\"true\" dtmfDetection=\"true\" welcomeGreeting=\"welcome_greeting\" partialPrompts=\"true\" interruptible=\"true\" interruptByDtmf=\"true\" welcomeGreetingInterruptible=\"true\" debug=\"true\"></ConversationRelay>" + Environment.NewLine +
+                "  <Assistant id=\"id\" language=\"language\" ttsLanguage=\"tts_language\" transcriptionLanguage=\"transcription_language\" ttsProvider=\"tts_provider\" voice=\"voice\" transcriptionProvider=\"transcription_provider\" speechModel=\"speech_model\" profanityFilter=\"true\" dtmfDetection=\"true\" welcomeGreeting=\"welcome_greeting\" partialPrompts=\"true\" interruptible=\"true\" interruptByDtmf=\"true\" welcomeGreetingInterruptible=\"true\" debug=\"true\"></Assistant>" + Environment.NewLine +
                 "</Connect>",
                 elem.ToString()
             );